node mcu displaying decibels on lcd changed so that the decibels are cast to uint8_t so that it fits on the lcd display

+++ node-mcu/src/main.cpp
@@ -133,7 +133,7 @@ void displayNoiseLevel(float decibels) {
   lcd.clear();
   lcd.setCursor(0, 0);
   lcd.print("Loudness: ");
-  lcd.print(decibels);
+  lcd.print((uint8_t) decibels);
   lcd.print("dB");
 
   if (decibels <= 60) {
